quelresto_backend/requirements.base.txt
Django==1.8.2
django-frontend-skeleton
django-extensions==1.5.5
djangorestframework==3.1.3
psycopg2-binary==2.8.3
six
